requirements.txt
Django==3.2.7
djangorestframework==3.12.4
Pillow==8.3.2